requirements.txt
streamlit
litellm
python-dotenv
pandas
fpdf2
pymupdf
flashtext
requests
tavily-python
